+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - 新增 `f-link` 组件 `beforeIcon` `afterIcon` 插槽
 - `f-avatar` 组件 `icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
 - 新增 `f-avatar` 组件 `icon` 插槽
+- 新增 `f-avatar` 组件 `icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
+- 废除 `f-avatar` 组件 `load-animation` 配置项
 
 - 废除 `f-loading` 组件 `icon` 配置项
 - 新增 `f-loading` 组件 `loadingIcon` 插槽